requirements.txt
backoff==2.2.1
openai==0.28.1
pdfminer.six==20221105
